+++ src/main/java/percistence/repositories/Interface/InterfaceEstudianteRepository.java
@@ -9,7 +9,7 @@ public interface InterfaceEstudianteRepository {
 
     public void addEstudiante(Estudiante p) throws Exception;
     public void updateEstudiante(Estudiante p, Estudiante pNew) throws Exception;
-    public void deleteEstudiante(Long id) throws Exception;
+    public void deleteEstudiante(int id) throws Exception;
     public List<Estudiante> getEstudiantes() throws Exception;
 
     public void addEstudianteToCarrera(Estudiante e, Carrera c) throws Exception;
